Refactored errors in DAL
