--- app/src/data_access_layer/errors/database.error.ts
@@ -1,9 +1,0 @@
-import { InternalServerError } from '../../common/errors';
-
-export class DatabaseError extends InternalServerError {
-  constructor(
-    message: string = 'An unexpected error occurred while accessing the database.',
-  ) {
-    super(message);
-  }
-}
